bug 4665: Make configure's PMI_HOME defaulting smarter for Cray PMI
Previously the defaulting algorithm for PMI_HOME/--with-pmi-home on Cray
systems had a hard requirement for the `cray-pmi` module to be loaded.

However this module is absent from the default environment of some Cray EX
systems, and we don't actually need it now that the Cray PMI
installer maintains a `/opt/cray/pe/pmi/default` symlink.

The prior logic trying to take advantage of this symlink was not working
as intended and was only active when the module was loaded, which is
exactly the case where it wasn't necessary.