Remove GASNET_SPAWN_CONTROL environment variable
This commit removes (almost) all use of the undocumented (internal)
`GASNET_SPAWN_CONTROL` environment variable, instead using the
user-facing `GASNET_<conduit>_SPAWNER` as the primary means to control
the logic in `gasneti_spawnerInit()`.

The only remaining use of `GASNET_SPAWN_CONTROL` is as an alternative
to `GASNET_<conduit>_SPAWNER` when the latter is not set.  This
provides support for use of an out-dated `gasnetrun_<conduit>` utility
(as risk mitigation, not as a recommended/endorsed use case).

In the absence of both `GASNET_<conduit>_SPAWNER` and
`GASNET_SPAWN_CONTROL`, the default is the conduit-specific value
established at configure time.  A significant fraction of this commit
is dedicated to making that configure-determined value a pre-processor
value while accommodating the fact that only three of seven current
conduits actually have/use such a default[1], but they all currently
compile `gasneti_spawnerInit()`.

Also notable is the removal of ofi-conduit's special-casing of the
Cray PMI case.  With the changes outlined above, the library now
contains logic to select the default `GASNET_OFI_SPAWNER` without
depending on the `gasnetrun_ofi` script.  Thus, the logic in
`configure` to select pmi-spawner as the default on a HPE Cray EX
system is now sufficient and no special case is required.

[1] The smp-, mpi- and udp-conduits don't call `gasneti_spawnerInit()`
at all, while aries-conduit calls it with `force_spawner="pmi"`.