ucx: cease use of personalized prefix for env vars
This commit resolves Bug 4670 "UCX environment personalized prefix not
working as we document" by simply not using the feature.

As documented in the bug report, UCX does not actually apply the
prefix as documented and its use was additionally causing
`GASNET_UCX_` to be a prefix for both UCX and ucx-conduit.