PMI support: detection on Cray EX w/o craype
On some Cray EX systems, `$CRAYPE_NETWORK_TARGET` may be unset at
configure time because there is no `craype-network-*` environment module
loaded.  For instance, this is the case when Intel's oneAPI compilers
are the default.

Since the existing configure option `--enable-force-arch=crayex` already
defaults the PMI version to `cray`, this commit ensures that the
Cray-specific probe logic is also enabled on this platform.