ofi: restore precedence of Cray PMI over MPI
This commit reverts one patch hunk from commit e766dd27 "Remove
GASNET_SPAWN_CONTROL environment variable" which had inadvertently
removed the intended special-case spawner precedence given to Cray
PMI (over MPI) in ofi-conduit.

In addition to restoring the forced-spawner logic, this commit updates
it to use the correct/current variable `GASNET_OFI_SPAWNER` as well as
logging a forced value as default for the environment variable (since
it is known to be unset or empty).