ibv: bug 4682 - rare failure of bogus assertion
This commit fixes bug 4682 by separating the non-NULL assertion from
the "maybe" refill of the spare reply buffer slot.  This eliminates
an `_IF()` suffixed macro in favor of `if()` guarding a call to the
unsuffixed macro.

In addition to removing the (rarely) incorrect assertion, the
placement of `gasnetc_maybe_restore_spare_reply_bbuf()` outside
the assertion ensures the intended behavior in an `NDEBUG` build
where previously only in a `DEBUG` build could a bounce buffer
released from a completed RMA operation be used to refill the
spare reply buffer slot.