ibv: refactor GASNETC_SND_REAP_COLLECT logic
This commit refactors the logic for `GASNETC_SND_REAP_COLLECT` to
enable "dynamic" enable/disable of the batching logic via an added
argument to the `COLLECT` and `FREE` family of macros.  This is
then used with a compile-time constant `1` or `0` to enable a single
inline function to be used inside a loop (with batching enabled) and
without a loop (without batching, to appear).