ibv: tweak for dynamically single-HCA runs
This commit adds branches to avoid thread contention on shared
counters when the counter values will always be `0`.

The use of shared counters for iterating over send and receive CQs of
distinct HCAs is skipped if a multi-rail configuration is running with
a single HCA.  Note that the shared counter logic is already
unreachable if multi-rail support was disable at configuration time.