PLPA: "safety" tweaks to some `PLPA_CPU_*` macros
This commit addresses issues with the `num` argument to the macros
`PLPA_CPU_{SET,CLR,ISSET}()`.

1. The expansion of `num` in the supporting `PLPA_CPU_{BYTE,BIT}()`
macros is parenthesized to avoid issues with an expression as an
argument and operator precedence with respect to `/` or `%`.

2. The macros `PLPA_CPU_{SET,CLR}()` now use a temporary variable to
prevent multiple evaluation of `num` due to calling both
`PLPA_CPU_BYTE()` and `PLPA_CPU_BIT()`.

3. `PLPA_CPU_ISSET()` is now a `static` function, giving multiple
evaluation protection via the function argument.