Add a perl-based GASNet ident script
This is a fully independent re-implementation of the ident utility
from the GNU RCS package that was written without any consultation
of the source files in that package.

The command-line interface and output mostly mimic the RCS package,
deviating only where that makes more sense.

It requires perl in the PATH.