More resilient logic for $BUILD_USER
As observed in bug 4698, it is possible for the command `id -un` to
produce output while also producing a non-zero exit code.  This
resulted in a `$BUILD_ID` string with an embedded newline.

This commit resolves the problem by ensuring that we discard the
output of `whoami` or `id -un` if they yield a non-zero exit code.