Chapel CI: do not build "developer mode" compiler
This commit ceases the practice of building Chapel's compiler with
`-Werror` in a debug-mode harness run.  Doing so has historically
meant periods in which our nightly debug-mode testing was blocked by
"new" fatal warnings in the compiler.  These might come from code
changes or new C/C++ compilers.  However, such things are not the
point of testing with GASNet-EX.

The runtime libraries are still built with debugging and assertions
enabled and `-Werror`.