Improve behaviour for certain RMA range checks
This commit changes the debug-only bounds checking logic to better
handle cases in which the row of `gasneti_seginfo_tbl[]` for the index
of the EP in question is missing.  `gasneti_client_seginfo()` will now
return a pointer to a "null segment" (all zero fields), allowing the
range check to fail in the same manner as would occur if the row were
present, but the entry were not populated.  Previously, the behaviour
was to fail an assertion that the row was present.