ibv: resolve bug 4617
This commit resolves Bug 4617: "SEGV in remote firehose, NDEBUG and
clang-16.0.0+" by adding a `volatile` qualifier as described in the
bug report.  The bug report also records results of testing with
clang from LLVM versions 15, 16, 17 and 18 which shows this commit
to be sufficient to resolve the observed failures of `make
run-tests-seq` in LARGE and EVERYTHING segment modes.

This change is believed to resolve the issue by inhibiting some
complier optimization(s) in clang from LLVM 16.0.0 or later.  However,
we do not know if the issue is a compiler bug or interaction between
valid optimizations and the extensive use of type-punning in firehose.

Note that the bug report documented two possible placements for
the `volatile` qualifier.  This commit applies only the first, in
part because the second leads to warnings.