mmap: fail gracefully if hugetlbfs is not usable
This commit updates the logic associated with runtime enable of
hugetlbfs support to fail with an informative error message if the
selected (possibly default) huge page size is not available.

This includes the important case in which the development package for
libhugetlbfs is present, but the actual system is not configured.
Previously, running on a system without hugetlbfs configured was very
likely to fail with "Bad file descriptor" errors when creating
segments.

Note: support for `--enable-pshm-hugetlbfs` remains experimental.