ofi: correct unsafe use of memcpy()
This commit replaces two erroneous instances of `memcpy()` with
`GASNETI_MEMCPY_SAFE_EMPTY()` in the "packed long" (a.k.a.
`OFI_AM_LONG_MEDIUM`) logic (one each in pack and unpack paths).

On both paths zero length with one "bad" address is possible with
the recent update to `testcore3`.  In such a case the behavior of bare
`memcpy()` is undefined.