ofi: convert last memcpy to GASNETI_MEMCPY
This commit replaces ofi-conduit's last remaining instance of
`memcpy()` with `GASNETI_MEMCPY()`.  This instance appears in the
logic to populate bounce-buffers and thus always has non-zero length
and disjoint addresses.  In _correct_ usage the addresses are both
valid.  While an unmapped `src` for a Put might propagate to this
point, handling that is not the purpose of the `GASNETI_MEMCPY*`
wrappers.