ibv: convert last memcpy to GASNETI_MEMCPY
This commit replaces ibv-conduit's last remaining instance of
`memcpy()` with `GASNETI_MEMCPY()`.  This instance appears in the XRC
setup logic and copies a table from the "leader" process to the
conduit's slice of the PSHM space to be shared by the other processes
in the same nbrhd.  The length is always non-zero with both addresses
valid and non-overlapping.