smp: convert last memcpy to GASNETI_MEMCPY
This commit replaces smp-conduit's last remaining instance of
`memcpy()` with `GASNETI_MEMCPY()`.  This instance appears in the
startup code and copies a table of PIDs from the "leader" process to
the conduit's slice of the PSHM space to be shared by the other
processes.  The length is always non-zero with both addresses valid
and non-overlapping.