ibv: simpler progress thread termination
This commit adds a `pthread_exit(NULL)` call near the end of the
role-agnostic progress function wrapper.

For the case in which `gex_Client_Init()` uses `pthread_create()` to
launch progress threads, this is indistinguishable from the `return
NULL` which has become unreachable.

In the case that a client is using `GEX_FLAG_DEFER_THREADS`, this
reduces two possible behaviors (return or exit) the client may
need to handle to just one (exit).  Documentation has already been
written with this new behavior in mind.