hwloc: implement GASNET_HWLOC_QUERY env var
Recent changes to the hwloc initialization logic switched the query of
the cpuset to be scoped to the calling thread, rather than the entire
process.  As explained in new text in the top-level `README` neither
thread nor process is perfect.  So, this commit implements a new
`GASNET_HWLOC_QUERY` environment variable to allow client code or an
end user to select their desired behavior.