ofi: fix LIST_DEVICES + kinds + generic provider
This commit restores proper operation of `GASNET_OFI_LIST_DEVICES=y`
when ofi-conduit has been configured with memory kinds support and the
"generic" provider.  In that case, the logic to list the available
devices was incorrectly excluding providers without HMEM support.

The fix is to defer setting the `FI_HMEM` bit in `hints->caps` until
after the optional device enumeration, as was already done for the
`FI_MR_HMEM` bit in the memory registration mode.