Fix a cosmetic error in GASNET_VERBOSEENV=1 output with long values
When the sum of the envvar key length and value length exceeded 55
characters, an overflow led to unintentionally passing a negative
field width to snprintf(), which ultimately had the effect of inserting
stray space characters at the end of the output line (in the field
where "(default)" normally appears), with the number of stray spaces
growing with the sum of the key/value length.

For very long envvar values, this resulted in a large number of space
characters being appended to the output line (which in turn have the
appearance of stray blank lines in a normal line-wrapped terminal).