ssh-spawner: update documentation
This commit revises a block comment of documentation in the
ssh-spawner logic to correctly reflect the current default
de-duplication of hostnames, which was previously described as
"planned" and "to do".

Additionally, the block comment has been spell checked.

Finally, two instances of "node" have been changed to either "host" or
"process", as appropriate.
Yes, the documentation previously used it both ways!