smp: gasnetrun_smp script
This commit adds a `gasnetrun_smp` script as well as an accompanying
`--with-ssh-spawner=...` configure option to set the default.  In
addition the `ssh`, `mpi` and `pmi`, this option accepts a `fork`
value to utilize the legacy spawner.

This includes the addition of smp-specific and fork-specific logic
to the generic `gasnetrun.pl` script.

The `fork-spawner` is the current default, to ensure that existing
launch logic in clients will "just work" unless a different spawner
is explicitly selected at configure or run time.