hwloc: use thread scope to query cpuset
This commit changes the hwloc initialization logic to query the cpuset
of the calling thread, rather than the entire process.  This is
believed to be less risky than the prior behavior due to the recent
observation that at least libfabric with the cxi provider expands the
cpuset of the process to encompass entire sockets (presumed to be due
to creation of a progress thread, though this has not been directly
observed).

Resolves bug 4734: "incorrect hwloc-based NIC selection with cxi
provider".