ofi: case-insensitive compare for provider names
This commit restores operation of `udp` provider for libfabric
versions 1.7 through 1.11, via use of case-insensitive string
comparison for all expression involving a provider name.

As noted in commit 87cdf2b0, the name "UDP" (all caps) was used prior
to 1.12.0 when the name became "udp".  While that commit fixed use of
the `udp` provider in a `--with-ofi-provider=generic` build of
ofi-conduit, an analogous issue has remained in ofi-conduit for the
case of `--with-ofi-provider=udp`.  Prior to this commit, one would
see failures like the following with libfabric 1.7 through 1.11:

```
*** FATAL ERROR (proc 0): in gasnetc_ofi_init() at [redacted]/ofi-conduit/gasnet_ofi.c:1186: OFI provider 'udp;ofi_rxd' selected at configure time is not available at run time.
```

Note that these changes include comparisons of the provider name
against constants such as "cxi" and "psm2".  While this is not
required to address the case of "UDP"-vs-"udp", it embodies a "best
practice" which libfabric appears to exercise as well.