UPCXX CI: extend "smoke test" to include install
This commit extends the `_upcxx_smoke_test` CI test to include the
UPC++ installation step, and to compile `upcxx_hello` (in both
thread modes) using the _installed_ `upcxx-meta`.

While the change to how `upcxx_hello` is built is *un*conditional,
this is not anticipated to have any negative impact on existing
testing.