NPAM: update `gasnetc_AM_Commit*` signature
This commit updates the signature of internal `gasnetc_AM_Commit*()`
functions to add a `commit_flags` argument and to change the return
type from `void` to `int`.  The macros in `gasnet_ammacros.h` to
implement the existing `gex_AM_Commit*()` APIs are revised to pass
zero for the new argument and to cast the result to `(void)`.

The net result is no change to any semantics.  However, the new
signature enables the implementation of `gex_AM_Commit*_v2()` without
code duplication.