NPAM: Commit*_v2/IMM support in reference impl
This commit implements non-trivial support for `commit_flags ==
GEX_FLAG_IMMEDIATE` in the reference implementation of
`gex_AM_Commit*_v2()`.

Specifically, when `GEX_FLAG_IMMEDIATE` is passed to the Commit*_v2()
call, it is passed to the underlying FPAM injection call and its
result is returned by the Commit call.