testcore2: seed random number generator
Prior to this commit the randomized `testcore2` did not make a call to
`TEST_SRAND()` to seed the PRNG, and consequently ran a deterministic
set of operations every time it was run.  Combined with the default
iteration count of only 30, this leaves some significant concern about
the completeness of coverage by this test.