tests: address pathological behaviour in LCG64
By default, `TEST_RAND()` and related macros in `tests/test.h` use a
64-bit linear congruential generator.  However, the `TEST_RAND()`
family has been using the least significant bits, which are known to
be less random than others.

It was only recently determined that the lowest 3 bits of the
generator have a cycle of length just eight.  In other words, they
cycle though all eight possible values in just eight consecutive
samples, and then repeat those eight *in the same order* indefinitely.
This was uncovered due to an extreme case in which this cycle was
observed to induce unbounded retries of failed injection with
`GEX_FLAG_IMMEDIATE` in `testcore2`, even though such an occurrence
would have vanishingly small probability with truly random numbers.

This commit addresses the issue sketched above by "mixing" (via XOR)
the high 32 bits into the low 32 bits of the generator's current value
for generation of results from `TEST_RAND()` (and family).  There are
no changes to the underlying generator; only to how its bits are used
each time it is sampled.