testexit: spilt seq/par ranges to ease maintenance
This commit splits the SEQ and PAR sub-tests of `testexit` to use
the respective ranges `[1,50]` and `[51,99]`.  This will enable
future work to add nearly 40 more tests to the SEQ set without
requiring any renumbering of the PAR set.

This commit also reduces the instances that a integer constant equal
to the `testid` is used (replacing them with `testid`, of course).
The intention of the range split is that no renumbering should ever
be needed; which means these replacements are unlikely to actually
reduce the future maintenance costs.  However, this seemed like the
"best practice" to use, and required very little added effort.