hwloc: stub recognition of TYPE=auto*
This commit documents and (trivially) implements "auto" as a "type"
for topology-aware environment variables.  Specifically, if the
type variable (`GASNET_IBV_PORTS_TYPE` or `GASNET_OFI_DEVICE_TYPE`)
has a value starting with "auto" (case insensitive), the entire
value is returned to the conduit _instead of_ a topology-aware
port/device selection.  In the future, this will trigger the
conduit to pass a device enumeration to additional hwloc logic
to complete the "automatic" device selection.

The top-level documentation defers to conduit-specific README for
the "auto" behavior, and in this "stub" implementation, both
ibv and ofi document "no support".