ibv: initial GASNET_IBV_PORTS_TYPE=auto support
This commit adds the logic in ibv-conduit necessary to utilize the new
`gasneti_hwloc_auto_select()` capability.

Due to limitations in the conduit-independent logic, only
single-HCA-per-proc is supported.  Due to limitations in the
conduit-specific logic, only `GASNET_IBV_PORTS` can be used to filter
(allow-list) HCAs.  Either or both of these limitations may be removed
in future work.

Adds corresponding user-facing documentation to the conduit `README`.