CI: remove TOP_{SRC,BUILD}DIR assumptions
This commit revises the `Makefile`s for each of the harness suites
embedded in the GASNet-EX sources to honor `$(GASNET_BLDDIR)` in
preference to `$(TOP_BUILDDIR)/gasnet`, and similarly for
`$(GASNET_SRCDIR)` and one instance of `$(UPCR_BLDDIR)`.  Recent
updates to the `pagoda-test-harness` export the new variables to the
harness environment, but these changes retain the old values relative
to `$(TOP_BUILDDIR)` and `$(TOP_SRCDIR)` as defaults if the needed
`GASNET` or `UPCR` variable is unset.

Additionally, use of `$(TOP_BUILDDIR)` as a default location for
product build directories has been replaced with use of
`$(HARNESS_WORKDIR)`.  This is safer, in general, but my have a
performance impact on systems where build directories are private and
work directories are shared.  In such cases, there are existing
variables one should set to override the default locations.

Overall, this commit addresses the "second half" of the issue
identified in bug 4745, regarding how the directory structure of
UPCR source and build directories had become hard-coded in multiple
places.