ofi: properly align buffer allocations
This commit resolves bug 4749 by ensuring that allocations of the
structure containing AM Medium send buffers is properly aligned. An
assertion is added to verify the proper alignment of the nested
structure field.

Other than the new assertion, this has no impact on an LP64 target
because `GASNETC_SIZEOF_AM_BUF_T` was already a multiple of eight.