ofi: fix "messy exits" on testexit NPAM subtests
This commit implements a new `gasneti_checknpam_disarm()` and deploys
it in ofi-conduit's `gasnetc_exit()` to prevent assertion failures in
the new exit-after-Prepare subtests of `testexit`.  This is done in
conduit-independent code both to maintain an abstraction boundary, and
because future implementations of native NPAM may need this as well.

Prior to this commit, a 4-process `testexit 14` would fail with the
following, and `testexit 15` similarly (with "Reply" in place of
"Request")

```
Running exit test 14: non-collective gasnet_exit(14) following PrepareRequestMedium()
*** FATAL ERROR (proc 3): in gasneti_checknpam() at /[redacted]/gasnet_am.c:579: Invalid GASNet call (communication injection or poll) between gex_AM_PrepareRequest() and the corresponding Commit on this thread
```