ratomic: fully honor ACQ fence in FCAS (bug #4754)
As described in detail in bug 4754, the tools-based implementation of
`GEX_AD_OP_FCAS` has not been honoring `GEX_FLAG_AD_ACQ` in the read
which determines the fetch value for the case no swap took place.
This commit corrects the `flags` argument to the atomic read operation
in two places (one each for integer and floating-point operands).

Resolves bug 4754: intermittent testratomic validation failures on
Apple M1 h/w since 2024.09.15