ibv: more intelligent multi-rail polling
This commit adds a `gasnetc_snd_reap_hca()` function which polls for
send completions on a specified HCA, rather than round-robin polling
(advanced across calls) as done in `gasnetc_snd_reap()`.

The logic in `gasnetc_snd_post_common()` to implement fenced puts is
modified to use the new function while polling for a SQ slot for the
required atomic operation.  Since it is necessary to allocate the SQ
slot for the atomic on the same HCA as the put, it makes more sense to
advance only the one specific HCA than to advance an arbitrary one on
each poll.  This may (marginally) improve performance of fenced puts.

Additionally, a typo ("CQ" where "SQ" was intended) has been fixed
in the block comment preceding the updated code.

In the absence of fenced puts, there are no behavioral changes in this
commit.  However, this change will enable future enhancements to the
implementation of `GEX_FLAG_IMMEDIATE`.