ibv: correct QP accounting in singleton runs
As reported in bug 4760, PR#386 changed the connection logic in
ibv-conduit to no longer skip intra-neighborhood connections but
failed to update logic for singleton (one process) runs with the
`GASNET_NUM_QPS` variable set to value larger than the default of `1`.

This commit removes the special casing of singleton runs which lead to
incorrect values for the `qps` and `max_qps` fields in the HCA data
structure.  That special case was intended to prevent division by zero
when the elision of intra-neighborhood would have prevented any
assignment to those fields.

Additionally updates the previously failing assertion to report the
values involved.