ibv: GASNET_IBV_PORTS_VERBOSE>1 prints HCAs
This commit enables logic to print the HCAs selected by each process,
previously reached only when `GASNET_SPAWN_VERBOSE` was true, when
`GASNET_IBV_PORTS_VERBOSE` is greater than one.  This variables was
already documented as integer (not Boolean) for purposes such as this.

Additionally, this commit moves the read of `GASNET_IBV_PORTS_VERBOSE`
from the environment to be slightly earlier, to enable future use of
the value in the implementation of `GASNET_IBV_PORTS_TYPE=auto`.