ibv: conditional run of progress funcs (bug 4764)
This commit resolves bug 4764 by preventing the problematic execution
of progress functions when polling for resources to send an AMRequest
header.  Other cases continue to run progress functions.

As noted in a new comment, this change is subject to possible
revision if we add a mechanism to distinguish non-communicating
progress functions (which *are* safe to run).