coll: fix set+enable race on gasnete_barrier_pf
This commit provides `GASNETI_PROGRESSFN_INITIALIZER` as the address
of a do-nothing progress function, and uses it as the initial value of
`gasnete_barrier_pf`.  This resolves the defect reported in Bug 4766
by ensuring that a thread observing an enable of the progress function
prior to observing its intended final value will run the new
do-nothing progress function (instead of a NULL pointer).

Resolves Bug 4766 - Executing a NULL pointer via gasnete_barrier_pf.