Legion CI: coverage for legion "Release" builds
This commit ensures that an explicit `DEBUG=0` is passed on the `make`
command line when building Legion tests with an `NDEBUG` build of
GASNet.  An explicit `DEBUG=1` for GASNet `DEBUG` builds protects
against any change to their default as well as the possibility that
`DEBUG` is set in the environment.