Fix gasneti_getenv_hwloc_withdefault() return type
This commit changes the type of `gasneti_getenv_hwloc_withdefault()`
from `char *` to `const char *`.  This reflects the fact that some
paths return strings from the environment, or string constants.

Additionally, the `const` qualifier is added to local variables in
the same function, and to appropriate variables in its two callers.