hwloc: conduits must opt-in to new "auto" support
This commit adds an `enable_auto` argument to
`gasneti_getenv_hwloc_withdefault()`, which is required to activate
the recently added recognition of "auto" as a special case.
Currently, ibv-conduit is the only conduit with the necessary support
and therefore is the only one to pass non-zero for the new argument.

This change ensures continuity with the behavior of ofi-conduit on the
`develop` branch and in previous releases.  Prior to this commit,
recent changes to `gasneti_getenv_hwloc_withdefault()` had
(unacceptably) made use of `GASNET_OFI_DEVICE_TYPE=auto` a fatal
error.