GASNet CI: update info collection via testhello
This commit makes two changes to how `gasnet-tests/testhello` is run,
including job layout and an ibv-specific environment variable:

    1. Now run with both forced 1-ppn (the way most of the `gasnet-tests`
       suite is run and the only way this test was run prior to this
       commit), as well as running with the harness's default PPN (the
       way most other test suites run).
    2. Now request the highest supported level of verbosity from
       ibv-conduit's device selection/enumeration.

Together these changes enable collection of data to evaluate the
effectiveness of `GASNET_IBV_PORTS_TYPE=auto` via the addition of
multi-ppn runs, without loss of the prior 1-ppn case which report the
configuration used in, for instance, the GASNet performance tests.

The addition of multi-ppn runs may also provide previously unavailable
information regarding PSHM.

Finally, the tests now carry a `-seq` suffix.  This was necessary when
creating a second test variant to avoid a need for new `Makefile` logic.
Adding the suffix to the "base" variant not only provides consistent
names, it also provides a fortuitous reduction in the likelihood of
comparing dissimilar runs with the same process count from before and
after the change to job layout.