Bug 4773 (partial): more informative error message
This commit addresses the "near-term" concern raised in bug 4773
regarding the quality of the fatalerror message issued when the
cross-mapping of PSHM segments fail.

Now the message looks like, for instance:

```
*** FATAL ERROR (proc 0): in gasneti_mmap_shared_internal() at /[redacted]/gasnet_mmap.c:927: cross-mapping failed for client segment of size 30842880: Cannot allocate memory
```

Where the message now distinguishes between allocation and
cross-mapping, and between client and aux segments.  The word
"segment" did not appear at all previously.