AMX: address pedantic warnings
While many (most?) modern compilers support it, the C preprocessor is
not required to support directives (such as `#if`) appearing within
arguments to a preprocessor macro.  Consequently, `-pedantic` may
warn about attempts to pass preprocessor-conditional arguments to
C library functions such as the `printf()` family, when they are
defined as macros.

This commit fixes such a call to `sprintf()` in AMMPI, using the simple
approach suggested in the C spec (and suggested to me by Dan) of adding
parenthesis around the function name.

Also addresses a similar call to `snprintf()` in AMPUDP, even though
that one has yet been seen to yield any warnings.