docs: close version 0.19 of the specification
This commit sets the advertised minor version of the GASNet-EX
specification to `19`, and advances the work-in-progress value of
`GEX_SPEC_VERSION_MINOR` in the specification document from `19` to
`20`.  Together these changes finalize/close version 0.19.

This recognizes the non-trivial changes surrounding the specification
and implementation of the `gex_AM_Commit*_v2()` and `gex_AM_Cancel*()`
API families.