template-conduit: update NPAM Medium example
This commit updates the template-conduit's example of FPAM Medium
injection (i.e. `gasnetc_AM{Request,Reply}MediumM()`) in a native-NPAM
build.  The examples now to pass `flags` to the commit step and act on
its return value.  This is "best practice" if the conduit is able to
return non-zero from the commit step, and matches the current ibv and
ofi implementations.