testvis: Fix a retry loop in PC arrival check
The VIS peer completion feature is currently implemented using a separate,
initiator-chained AM Medium to deliver the PC payload and trigger the
handler execution at the target.

testvis includes a "peer completion arrival" check designed to validate the
number of PC requests sent by the initiator match those delivered to the
target. It does this using a barrier to establish VIS operation quiescence
and a put to transmit the expected count from initiator to target.

The interface provides no way for the VIS initiator to "drain" delivery of PC
handlers to the remote side (and we have no way in general to fence AMs),
so there's a potential race on conduits with unordered p2p AM where the AM
for a PC might take a "long path" and not arrive early enough to be included
in the validation performed at the target. testvis attempts to resolve this
acknowledged race by inserting up to a 5 second delay at the target while
awaiting arrival of any "straggler" PC handlers. However due to an oversight,
that delay loop failed to include an AM Poll operation, which of course is
required for any straggler PC handler to actually be dispatched and counted.

On conduits with unordered AM delivery (e.g. udp-conduit with packet loss)
this was causing rare intermittent spurious failures of the form:
```
ERROR: node 2/4 peer completion arrival mismatch expect=46 actual=44
```