Expand support for compiler-based membars
Currently (both before and after this commit), `gasnet_membar.h` uses
`__sync_synchronize()` for all memory fences on the RISC-V, Tilera and
IBM System 390 (SystemZ) architectures.  What this commit changes is
how and when that implementation is selected.

Prior to this commit, this implementation of memory barriers was
available only when using a GNU compiler.  Consequently, even though
LLVM Clang claims support for RISC-V and SystemZ, we could not use
`clang` on either system (library compilation would fail with a
`#error` in `gasnet_membar.h` about an "unknown compiler").

To enable use of `clang` on RISC-V, this commit goes beyond just
appending `|| PLATFORM_COMPILER_CLANG` after `PLATFORM_COMPILER_GNU`,
and instead introduces a probe in `configure` to detect the desired
support.  This feature-based approach is "the autoconf way".

Verified on RISC-V with both `gcc`/`g++` and `clang`/`clang++`.