fixup - UPCXX CI: expand K/F for allocator-example
The commit to allow for a KnowFailure of `allocator-example` with the
LLVM/clang compilers failed to recognize that LLVM's `libc++` *also*
fails in "opt" codemode (with an illegal instruction and no helpful
message).  Instrumentation at the source code level has confirmed that
the failure is on the same line of code as with "debug" codemode.