fixup - UPCXX CI: rework environment handling
This commit corrects the erroneous removal of "UNSET_VARS" logic in
early 2020 when the handling of environment variables in
`external-upcxx/Makefile` was simplified in commit e96e2517f.

The original use of the `$(UNSET_VARS)` variable was to ensure that
when GASNet is configured via UPC++'s `configure` (as opposed to prior
to execution of the test harness), the environment exported _by_ the
harness did not pollute the configure of GASNet.  This was lost,
resulting in (at least) anomalous "doubling" of `CFLAGS` and
`CXXFLAGS` when `UPCXX_GASNET` was set to either a tarball URL or a
GASNet source directory.

The doubled flags can be seen in the history of the
`EX-dirac-smp-gcc-pshm` pushbuild configuration (until its retirement
in early September 2024), because it set `UPCXX_GASNET` to the
harness's GASNet source directory.

It is worth noting that, due to this unintended regression on early
2020, pushbuild configuration `EX-cgpu-smp_nopshm-gcc` (which also
sets `UPCXX_GASNET` to the harness's GASNet source directory)
currently includes extra effort to override the harness-provided
environment via higher-precedence `--with-*=` flags in
`UPCXX_CONFIGURE_EXTRA`.

This commit restores the original intent, by using a renamed
`$(UNSET_HARNESS_VARS)` in both the UPC++ configure and build steps.
This is required in the configure step because (at least) `CFLAGS`,
`CXXFLAGS` and `LDFLAGS` are captured for later replay to GASNet's
configure in the build step. Because `CPPFLAGS` and `LIBS` are _not_
captured in UPC++'s configure step, they must be cleared for the build
step to prevent GASNet's configure step from consuming them.  In the
event that either are necessary on a given system, one can use
`--with-cppflags` or `--with-libs` in `UPCXX_CONFIGURE_EXTRA`.  Note
that since we unset the _entire_ set of variables for _both_ steps,
the new logic will remain effective if there are changes to the
details given above.