LEGION CI: work arounds for current Realm status
This commit introduces fixes for two issues which are present in the
'stable' branch of Legion and not in its 'master' branch.  Both exist
within the `realm/` directory, which is "in flux" as the maintainers
move to separate the Legion and Realm products.

Both work-arounds have been tested against the current 'master` and
found not to have negative impact.